fix: move ParticleWithSpin to particle
