+++ src/expertsystem/amplitude/helicity.py
@@ -16,8 +16,12 @@ from expertsystem.reaction import Result
 from expertsystem.reaction.combinatorics import (
     perform_external_edge_identical_particle_combinatorics,
 )
-from expertsystem.reaction.particle import Particle, ParticleCollection, Spin
-from expertsystem.reaction.quantum_numbers import ParticleWithSpin
+from expertsystem.reaction.particle import (
+    Particle,
+    ParticleCollection,
+    ParticleWithSpin,
+    Spin,
+)
 from expertsystem.reaction.topology import StateTransitionGraph
 
 from ._graph_info import (
